Sprint 3 Machine Learning y Analitics/StreamLit_Sistema_Recomendacion/requirements.txt
streamlit
pandas
plotly
scikit-learn
db-dtypes
google-cloud-bigquery
numpy
wordcloud
regex
Unidecode
nltk
matplotlib
textblob